Modif after PSScriptAnalyzer

+++ lib/Scripts/Remove-UserProxy.ps1
@@ -11,18 +11,22 @@ function Remove-UserProxy {
             Purpose/Change: Initial script development
 
 		.EXAMPLE 
-            Get-Proxy
+            Remove-UserProxy
         
 	#>
 
-    [CmdletBinding()] 
+    [CmdletBinding(SupportsShouldProcess)] 
     Param () 
  
     Begin {
 
-        # Set VerbosePreference to Continue so that verbose messages are displayed. 
-        $VerbosePreference = 'Continue'
         $regKey = "HKCU:\Software\Microsoft\Windows\CurrentVersion\Internet Settings"
+		if (-not $PSBoundParameters.ContainsKey('Confirm')) {
+            $ConfirmPreference = $PSCmdlet.SessionState.PSVariable.GetValue('ConfirmPreference')
+        }
+        if (-not $PSBoundParameters.ContainsKey('WhatIf')) {
+            $WhatIfPreference = $PSCmdlet.SessionState.PSVariable.GetValue('WhatIfPreference')
+        }
     } 
     
     Process {
@@ -31,23 +35,29 @@ function Remove-UserProxy {
         
         if ($ProxyInfo.AutoConfigURL) {
             Write-Verbose "PROXY CONFIG : $($ProxyInfo.AutoConfigURL)"
-            Remove-ItemProperty -Path $regkey -Name AutoConfigURL -force
-             if ($ProxyInfo.ProxyServer) {
-                Remove-ItemProperty -Path $regkey -Name ProxyServer -force -EA 0
-                Set-ItemProperty -Path $regkey -Name ProxyEnable -Value 0
-            }
+            if ($pscmdlet.shouldprocess("Are you sure?")) {
+				Remove-ItemProperty -Path $regkey -Name AutoConfigURL -force
+				if ($ProxyInfo.ProxyServer) {
+					Remove-ItemProperty -Path $regkey -Name ProxyServer -force -EA 0
+					Set-ItemProperty -Path $regkey -Name ProxyEnable -Value 0
+				}
+				Write-Verbose "PROXY CONFIG : removed"
+			}
         }
         elseif ($ProxyInfo.ProxyEnable -eq 1) {
-             Write-Verbose "PROXY CONFIG : $($ProxyInfo.ProxyServer)"
-             Remove-ItemProperty -Path $regkey -Name ProxyServer -force -EA 0
-             Set-ItemProperty -Path $regkey -Name ProxyEnable -Value 0
-             Set-ItemProperty -Path $regkey -Name ProxyOverride -Value $(($ProxyInfo.ProxyOverride).replace(";<local>",""))
+			Write-Verbose "PROXY CONFIG : $($ProxyInfo.ProxyServer)"
+			if ($pscmdlet.shouldprocess("Are you sure?")) { 
+				Remove-ItemProperty -Path $regkey -Name ProxyServer -force -EA 0
+				Set-ItemProperty -Path $regkey -Name ProxyEnable -Value 0
+				Set-ItemProperty -Path $regkey -Name ProxyOverride -Value $(($ProxyInfo.ProxyOverride).replace(";<local>",""))
+			}
+			Write-Verbose "PROXY CONFIG : removed"
         }
         else {
-            Write-host "Proxy Not enable"
+            Write-verbose "Proxy Not enable"
        }
     } 
     
     End {} 
-    
+
 }
